% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{correct_naming}
\alias{correct_naming}
\title{Check column names}
\usage{
correct_naming(tab, type = "raw")
}
\arguments{
\item{tab}{Tibble to check}

\item{type}{One of "raw", "eval", "score"}
}
\value{
Logical indicating whether the column names are correct.
}
\description{
Checks if the column names are compatible with the processing functions
(\code{evaluate_raw_data}, \code{evaluate_scores}).
}
\details{
Depending on the \code{type} value, it checks
\describe{
\item{raw}{Can be input to \code{evaluate_raw_data}}
\item{eval}{Can be input to \code{evaluate_scores}, can be output of \code{evaluate_raw_data}}
\item{score}{can be output of \code{evaluate_scores}}
...
}
}
\examples{
\dontrun{
data("vfq_example")
correct_naming(vfq_example, "raw")
}
}
