% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{evaluate_raw_data}
\alias{evaluate_raw_data}
\title{Evaluate VFQ raw data}
\usage{
evaluate_raw_data(tab)
}
\arguments{
\item{tab}{Tibble with correctly named columns}
}
\value{
Tibble with columns
\code{sVFQ01} to \code{sVFQ25} with values from 0 to 100.
Each row corresponds to one row in the input \code{tab}.
}
\description{
Adjusts scores for each item
(they differ based on the number of options).
}
