% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{evaluate_scores}
\alias{evaluate_scores}
\title{Calculates VFQ scale scores}
\usage{
evaluate_scores(stab)
}
\arguments{
\item{stab}{Tibble obtained via \code{evaluate_raw_data}}
}
\value{
Tibble with 12 columns (\code{general_health} to \code{peripheral_vision})
}
\description{
Calculates VFQ scale scores
}
