% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vfq_example}
\alias{vfq_example}
\title{Random VFQ data}
\format{
\subsection{\code{who}}{

A tibble with 5 rows and 30 columns:
\describe{
\item{id}{unique id}
\item{VFQ01}{First item}
\item{VFQ02}{Second item}
...
}
}
}
\usage{
vfq_example
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
